extract out colors.config
